% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManagement.R
\name{cbindRecusive}
\alias{cbindRecusive}
\title{Recursively rbind a named list}
\usage{
cbindRecusive(Y, struct)
}
\arguments{
\item{Y}{\code{list}, list}

\item{struct}{\code{character}, variables names}
}
\description{
Recursively rbind a named list
}
