% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setsimulationPathH5.R
\name{getOptionsH5}
\alias{getOptionsH5}
\title{Get simulation path for h5 file}
\usage{
getOptionsH5(path)
}
\arguments{
\item{path}{\code{character} path of h5 file}
}
\description{
Get simulation path for h5 file
}
