% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeAttrib.R
\name{giveFormat}
\alias{giveFormat}
\title{Rescover format after read attributes from h5 file.}
\usage{
giveFormat(attribList)
}
\arguments{
\item{attribList}{\code{list}, list of study attributes read from h5.}
}
\description{
Rescover format after read attributes from h5 file.
}
