% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antaresReadH5.R
\name{h5ReadAntares}
\alias{h5ReadAntares}
\title{Read data}
\usage{
h5ReadAntares(path, areas = NULL, links = NULL, clusters = NULL,
  districts = NULL, mcYears = NULL, timeStep = "hourly", select = "all",
  showProgress = TRUE, simplify = TRUE, perf = TRUE)
}
\arguments{
\item{path}{{character} path of h5file to load}

\item{areas}{see \code{\link{antaresRead::readAntares}}}

\item{links}{see \code{\link{antaresRead::readAntares}}}

\item{clusters}{see \code{\link{antaresRead::readAntares}}}

\item{districts}{see \code{\link{antaresRead::readAntares}}}

\item{mcYears}{see \code{\link{antaresRead::readAntares}}}

\item{timeStep}{see \code{\link{antaresRead::readAntares}}}

\item{select}{see \code{\link{antaresRead::readAntares}}}

\item{showProgress}{see \code{\link{antaresRead::readAntares}}}

\item{simplify}{see \code{\link{antaresRead::readAntares}}}

\item{perf}{\code{boolean}, eval performance during developpement time, to remove}
}
\description{
Read data
}
