% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readH5.R
\name{readAntaresH5}
\alias{readAntaresH5}
\title{Read the data of an Antares simulation}
\usage{
readAntaresH5(path, areas = NULL, links = NULL, clusters = NULL,
  districts = NULL, mcYears = NULL, timeStep = "hourly", select = "all",
  showProgress = TRUE)
}
\arguments{
\item{path}{\code{character}, path of h5 file to read.}

\item{areas}{Vector containing the names of the areas to import. If
\code{NULL} no area is imported. The special value \code{"all"} tells the
function to import all areas. By default, the value is "all" when no other argument is enter and "NULL" when other arguments are enter.}

\item{links}{Vector containing the name of links to import. If \code{NULL} no
area is imported. The special value \code{"all"} tells the function to
import all areas. Use function \code{\link{antaresRead::getLinks}} to import all links
connected to some areas.}

\item{clusters}{Vector containing the name of the areas for which you want to
import results at cluster level. If \code{NULL} no cluster is imported. The
special value \code{"all"} tells the function to import clusters from all
areas.}

\item{districts}{Vector containing the names of the districts to import. If \code{NULL},
no district is imported. The special value \code{"all"} tells the function to import all
districts.}

\item{mcYears}{Index of the Monte-Carlo years to import. If \code{NULL}, synthetic results
are read, else the specified Monte-Carlo simulations are imported. The
special value \code{all} tells the function to import all Monte-Carlo
simulations.}

\item{timeStep}{Resolution of the data to import: hourly (default), daily,
weekly, monthly or annual.}

\item{select}{Character vector containing the name of the columns to import. If this
argument is \code{NULL}, all variables are imported. Special names
\code{"allAreas"} and \code{"allLinks"} indicate to the function to import
all variables for areas or for links.}

\item{showProgress}{If TRUE the function displays information about the progress of the
importation.}
}
\value{
An object of class "antaresDataList" is returned. It is a list of
data.tables, each element representing one type of element (areas, links,
clusters)
}
\description{
\code{\link{readAntaresH5}} Has the same behavior than \code{\link{antaresRead::readAntares}} but it based on h5 file.
}
