% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setsimulationPathH5.R
\name{setSimulationPathH5}
\alias{setSimulationPathH5}
\title{Set simulation path for h5 file}
\usage{
setSimulationPathH5(path)
}
\arguments{
\item{path}{\code{character} path of h5 file}
}
\description{
Set simulation path for h5 file
}
