% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformData.R
\name{transformH5}
\alias{transformH5}
\title{Transform a \code{antaresDataList} object to be writable by \link{writeAntaresH5}}
\usage{
transformH5(data, areasKey = c("area"), linksKey = c("link"),
  districtKey = c("district"), clustersKey = c("area", "cluster"))
}
\arguments{
\item{data}{\code{antaresDataList}}

\item{areasKey}{\code{character} organization key for areas, define h5 group and subgroup}

\item{linksKey}{\code{character} organization key for links, define h5 group and subgroup}

\item{districtKey}{\code{character} organization key for districts, define h5 group and subgroup}

\item{clustersKey}{\code{character} organization key for clusters, define h5 group and subgroup}
}
\description{
Transform a \code{antaresDataList} object to be writable by \link{writeAntaresH5}
}
