% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeData.R
\name{writeAntaresData}
\alias{writeAntaresData}
\title{Write data}
\usage{
writeAntaresData(data, path, rootGroup = NULL, writeStructure = TRUE,
  mcAll = FALSE, compress = 0)
}
\arguments{
\item{data}{\code{antaresDataList}}

\item{path}{\code{character} patch of h5 file}

\item{rootGroup}{\code{character} group will contain all h5 organization}

\item{writeStructure}{\code{boolean}, write group and subgroup (only for first MCyear)}

\item{mcAll}{\code{character}, write mcAll}

\item{compress}{\code{numeric}, compression level}
}
\description{
Write data
}
