% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeH5ByYear.R
\name{writeAntaresH5}
\alias{writeAntaresH5}
\title{Convert antares output to h5 file}
\usage{
writeAntaresH5(path, timeSteps = c("hourly", "daily", "weekly", "monthly",
  "annual"), opts = antaresRead::simOptions(), writeMcAll = TRUE,
  compress = 1, misc = FALSE, thermalAvailabilities = FALSE,
  hydroStorage = FALSE, hydroStorageMaxPower = FALSE, reserve = FALSE,
  linkCapacity = FALSE, mustRun = FALSE, thermalModulation = FALSE)
}
\arguments{
\item{path}{\code{character} path of h5 file to write}

\item{timeSteps}{\code{character} timeSteps}

\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}. Defaut to \code{antaresRead::simOptions()}}

\item{writeMcAll}{\code{boolean} write mc-all}

\item{compress}{\code{numeric} compress level}

\item{misc}{\code{boolean} see \link[antaresRead]{readAntares}}

\item{thermalAvailabilities}{\code{boolean} see \link[antaresRead]{readAntares}}

\item{hydroStorage}{\code{boolean} see \link[antaresRead]{readAntares}}

\item{hydroStorageMaxPower}{\code{boolean} see \link[antaresRead]{readAntares}}

\item{reserve}{\code{boolean} see \link[antaresRead]{readAntares}}

\item{linkCapacity}{\code{boolean} see \link[antaresRead]{readAntares}}

\item{mustRun}{\code{boolean} see \link[antaresRead]{readAntares}}

\item{thermalModulation}{\code{boolean} see \link[antaresRead]{readAntares}}
}
\description{
Convert antares output to h5 file
}
