% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeAttrib.R
\name{writeAttribAndCreatGroup}
\alias{writeAttribAndCreatGroup}
\title{Write list from attributes in h5.}
\usage{
writeAttribAndCreatGroup(path, Y, group = NULL)
}
\arguments{
\item{Y}{\code{list}, list of study attributes}

\item{group}{\code{character} group where attributes will be write}
}
\description{
Write list from attributes in h5.
}
