% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeData.R
\name{writeDataType}
\alias{writeDataType}
\title{Write data by type}
\usage{
writeDataType(data, path, type, rootGroup = NULL, writeStructure = TRUE,
  mcAll = FALSE, compress = 0)
}
\arguments{
\item{data}{\code{antaresDataList}}

\item{path}{\code{character} patch of h5 file}

\item{type}{\code{character} type of data to write, must be areas, links, districts or clusters}

\item{rootGroup}{\code{character} group will contain all h5 organization}

\item{writeStructure}{\code{boolean}, write group and subgroup (only for first MCyear)}

\item{mcAll}{\code{character}, write mcAll}

\item{compress}{\code{boolean}, compress level}
}
\description{
Write data by type
}
