% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeAttrib.R
\name{writeList}
\alias{writeList}
\title{Write list from attributes in h5.}
\usage{
writeList(path, Y, group = NULL)
}
\arguments{
\item{path}{\code{character} patch of h5 file}

\item{Y}{\code{list}, list of study attributes}

\item{group}{\code{character} group where attributes will be write}
}
\description{
Write list from attributes in h5.
}
