% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeManagement.R
\name{writeTime}
\alias{writeTime}
\title{Write time in h5 file}
\usage{
writeTime(data, path, group)
}
\arguments{
\item{data}{\code{antaresDataList} see \link{readAntares}}

\item{path}{\code{character} path of h5 file}

\item{group}{\code{group} group where time are stocked}
}
\description{
Write time in h5 file
}
