% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeData.R
\name{creatGroup}
\alias{creatGroup}
\title{Create a group organization which subgroup.}
\usage{
creatGroup(groupIn = NULL, groupData, path)
}
\arguments{
\item{groupIn}{\code{character} name of main group (who will contain all organization)}

\item{groupData}{\code{data.table} organization}

\item{path}{\code{character} patch of h5 file}
}
\description{
Create a group organization which subgroup.
}
