% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeManagement.R
\name{getAllDateInfoFromDate}
\alias{getAllDateInfoFromDate}
\title{Read time and generate column who can be calculate from time}
\usage{
getAllDateInfoFromDate(path, group)
}
\arguments{
\item{path}{\code{character} path of h5 file}

\item{group}{\code{group} group where time are stocked}
}
\description{
Read time and generate column who can be calculate from time
}
