% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeManagement.R
\name{getAllDateInfoFromDate}
\alias{getAllDateInfoFromDate}
\title{Read time and generate column who can be calculate from time}
\usage{
getAllDateInfoFromDate(fid, group)
}
\arguments{
\item{fid}{\code{H5IdComponent} id of h5 file open which \link[rhdf5]{H5Fopen}}

\item{group}{\code{group} group where time are stocked}
}
\description{
Read time and generate column who can be calculate from time
}
