% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeManagement.R
\name{getTime}
\alias{getTime}
\title{Get timeId from antares study}
\usage{
getTime(data, timeStep)
}
\arguments{
\item{data}{\code{antaresDataList} see \link{readAntares}}

\item{timeStep}{\code{character} timeStep}
}
\description{
Get timeId from antares study
}
