% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antaresReadH5.R
\name{h5ReadAntares}
\alias{h5ReadAntares}
\title{Read data}
\usage{
h5ReadAntares(path, areas = NULL, links = NULL, clusters = NULL,
  districts = NULL, mcYears = NULL, timeStep = "hourly", select = "all",
  showProgress = TRUE, simplify = TRUE)
}
\description{
Read data
}
