% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setsimulationPathH5.R
\name{setSimulationPathH5}
\alias{setSimulationPathH5}
\title{Set simulation path for h5 file}
\usage{
setSimulationPathH5(path, simulation = NULL)
}
\arguments{
\item{path}{\code{character} path of a .h5 file or a repertory with .h5 file(s)}

\item{simulation}{\code{character} simulation number or name}
}
\description{
Set simulation path for h5 file
}
