% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeData.R
\name{writeAntaresData}
\alias{writeAntaresData}
\title{Write a h5 file from an object load with \link{readAntares}}
\usage{
writeAntaresData(data, path, rootGroup = NULL, writeStructure = TRUE,
  writeMCallName = FALSE, compress = 0)
}
\arguments{
\item{data}{\code{antaresDataList}}

\item{path}{\code{character} patch of h5 file}

\item{rootGroup}{\code{character} group will contain all h5 organization}

\item{writeStructure}{\code{boolean}, write group and subgroup (only for first MCyear)}

\item{writeMCallName}{\code{character}, write mc-all names}
}
\description{
Write a h5 file from an object load with \link{readAntares}
}
