% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeData.R
\name{writeAntaresH5}
\alias{writeAntaresH5}
\title{Write a h5 file from an object load with \link{readAntares}}
\usage{
writeAntaresH5(data, path, rootGroup = NULL)
}
\arguments{
\item{data}{\code{antaresDataList}}

\item{path}{\code{character} patch of h5 file}

\item{rootGroup}{\code{character} group will contain all h5 organization}
}
\description{
Write a h5 file from an object load with \link{readAntares}
}
